pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dalong-catboost"
version = "0.1.0"
description = "CatBoost project scaffold for training and inference"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Dalong" }
]
dependencies = [
  "catboost>=1.2.5",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "black>=24.3.0",
  "isort>=5.13.0",
  "ruff>=0.4.0",
  "mypy>=1.8.0",
  "ipykernel>=6.29.0",
]

[project.scripts]
dalong-catboost = "dalong_catboost.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["dalong_catboost*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]


